Moved transfer logic to separate util class.
