+++ SitecoreData.DataProviders.MongoDB.Tests/DataTransferTests.cs
@@ -53,7 +53,7 @@ namespace SitecoreData.DataProviders.MongoDB.Tests
         {
             var sourceDatabase = Factory.GetDatabase("master");
             var targetDatabase = Factory.GetDatabase("nosqlmongotest");
-            TransferPath("/sitecore/layout", sourceDatabase, targetDatabase);
+            TransferUtil.TransferPath("/sitecore/layout", sourceDatabase, targetDatabase);
             Assert.That(_db.GetCollectionNames().Contains("items"), Is.True);
         }
 
@@ -62,7 +62,7 @@ namespace SitecoreData.DataProviders.MongoDB.Tests
         {
             var sourceDatabase = Factory.GetDatabase("master");
             var targetDatabase = Factory.GetDatabase("nosqlmongotest");
-            TransferPath("/sitecore/layout", sourceDatabase, targetDatabase);
+            TransferUtil.TransferPath("/sitecore/layout", sourceDatabase, targetDatabase);
             Assert.That(_db.GetCollection("items").Count(), Is.EqualTo(59));
         }
 
@@ -71,83 +71,9 @@ namespace SitecoreData.DataProviders.MongoDB.Tests
         {
             var sourceDatabase = Factory.GetDatabase("master");
             var targetDatabase = Factory.GetDatabase("nosqlmongotest");
-            TransferPath("/sitecore/layout/renderings", sourceDatabase, targetDatabase);
+            TransferUtil.TransferPath("/sitecore/layout/renderings", sourceDatabase, targetDatabase);
             Assert.That(_db.GetCollection("items").Count(), Is.EqualTo(13));
 
-        }
-
-        private void TransferPath(string itemPath, Database sourceDatabase, Database targetDatabase)
-        {
-            var item = sourceDatabase.GetItem(itemPath);
-            var dataProvider = targetDatabase.GetDataProviders().First() as DataProviderWrapper;
-            TransferAncestors(item.Parent, dataProvider);
-            TransferItemAndDescendants(item, dataProvider);
-        }
-
-        //TODO Move to separate class, and refactor Transfer.aspx to use this new class.
-        public void TransferItemAndDescendants(Item item, DataProviderWrapper provider)
-        {
-            
-            TransferSingleItem(item, provider);
-
-            if (!item.HasChildren)
-            {
-                return;
-            }
-
-            foreach (Item child in item.Children)
-            {
-                TransferItemAndDescendants(child, provider);
-            }
-        }
-        
-        public void TransferAncestors(Item item, DataProviderWrapper provider)
-        {
-            if (item == null) return;
-            TransferAncestors(item.Parent,provider);
-            TransferSingleItem(item, provider);
-        }
-
-        private static void TransferSingleItem(Item item, DataProviderWrapper provider)
-        {
-            ItemDefinition parentDefinition = null;
-
-            if (item.Parent != null)
-            {
-                parentDefinition = new ItemDefinition(item.Parent.ID, item.Parent.Name, item.Parent.TemplateID,
-                                                      item.Parent.BranchId);
-            }
-
-            // Create the item in database
-            if (provider.CreateItem(item.ID, item.Name, item.TemplateID, parentDefinition, null))
-            {
-                foreach (var language in item.Languages)
-                {
-                    using (new LanguageSwitcher(language))
-                    {
-                        var itemInLanguage = item.Database.GetItem(item.ID);
-
-                        if (itemInLanguage != null)
-                        {
-                            // Add a version
-                            var itemDefinition = provider.GetItemDefinition(itemInLanguage.ID, null);
-
-                            // TODO: Add all version and not just v1
-                            provider.AddVersion(itemDefinition, new VersionUri(language, Sitecore.Data.Version.First), null);
-
-                            // Send the field values to the provider
-                            var changes = new ItemChanges(itemInLanguage);
-
-                            foreach (Field field in itemInLanguage.Fields)
-                            {
-                                changes.FieldChanges[field.ID] = new FieldChange(field, field.Value);
-                            }
-
-                            provider.SaveItem(itemDefinition, changes, null);
-                        }
-                    }
-                }
-            }
         }
 
         [Test]
