+++ src/SitecoreData.DataProviders.MongoDB.Tests/SitecoreData.DataProviders.MongoDB.Tests.csproj
@@ -71,6 +71,7 @@
     <Compile Include="Test.aspx.designer.cs">
       <DependentUpon>Test.aspx</DependentUpon>
     </Compile>
+    <Compile Include="TransferUtil.cs" />
     <Compile Include="UsingFastQueryWithMongo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
